requirements-dev.txt
pytest==5.3.0
pytest-cov==2.8.1
pytest-django==3.7.0
pytest-pythonpath==0.7.3
pytest-sugar==0.9.2
